Update schema and user query

+++ db/migration/000001_init_schema.up.sql
@@ -12,7 +12,7 @@ CREATE TABLE "users" (
 );
 
 CREATE TABLE "posts" (
-  "id" integer PRIMARY KEY,
+  "id" bigserial PRIMARY KEY,
   "title" varchar NOT NULL,
   "body" text,
   "user_id" integer NOT NULL,
